// src/api/gql/graphql.ts
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any }
  /** The `Money` scalar type represents monetary values and supports signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Money: { input: any; output: any }
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: any; output: any }
}
// ...
export type CurrentUser = {
  __typename?: "CurrentUser"
  channels: Array<CurrentUserChannel>
  id: Scalars["ID"]["output"]
  identifier: Scalars["String"]["output"]
}

export type CurrentUserChannel = {
  __typename?: "CurrentUserChannel"
  code: Scalars["String"]["output"]
  id: Scalars["ID"]["output"]
  permissions: Array<Permission>
  token: Scalars["String"]["output"]
}
// ...
export enum ErrorCode {
  AlreadyLoggedInError = "ALREADY_LOGGED_IN_ERROR",
  CouponCodeExpiredError = "COUPON_CODE_EXPIRED_ERROR",
  CouponCodeInvalidError = "COUPON_CODE_INVALID_ERROR",
  CouponCodeLimitError = "COUPON_CODE_LIMIT_ERROR",
  EmailAddressConflictError = "EMAIL_ADDRESS_CONFLICT_ERROR",
  GuestCheckoutError = "GUEST_CHECKOUT_ERROR",
  IdentifierChangeTokenExpiredError = "IDENTIFIER_CHANGE_TOKEN_EXPIRED_ERROR",
  IdentifierChangeTokenInvalidError = "IDENTIFIER_CHANGE_TOKEN_INVALID_ERROR",
  IneligiblePaymentMethodError = "INELIGIBLE_PAYMENT_METHOD_ERROR",
  IneligibleShippingMethodError = "INELIGIBLE_SHIPPING_METHOD_ERROR",
  InsufficientStockError = "INSUFFICIENT_STOCK_ERROR",
  InvalidCredentialsError = "INVALID_CREDENTIALS_ERROR",
  MissingPasswordError = "MISSING_PASSWORD_ERROR",
  NativeAuthStrategyError = "NATIVE_AUTH_STRATEGY_ERROR",
  NegativeQuantityError = "NEGATIVE_QUANTITY_ERROR",
  NotVerifiedError = "NOT_VERIFIED_ERROR",
  NoActiveOrderError = "NO_ACTIVE_ORDER_ERROR",
  OrderLimitError = "ORDER_LIMIT_ERROR",
  OrderModificationError = "ORDER_MODIFICATION_ERROR",
  OrderPaymentStateError = "ORDER_PAYMENT_STATE_ERROR",
  OrderStateTransitionError = "ORDER_STATE_TRANSITION_ERROR",
  PasswordAlreadySetError = "PASSWORD_ALREADY_SET_ERROR",
  PasswordResetTokenExpiredError = "PASSWORD_RESET_TOKEN_EXPIRED_ERROR",
  PasswordResetTokenInvalidError = "PASSWORD_RESET_TOKEN_INVALID_ERROR",
  PasswordValidationError = "PASSWORD_VALIDATION_ERROR",
  PaymentDeclinedError = "PAYMENT_DECLINED_ERROR",
  PaymentFailedError = "PAYMENT_FAILED_ERROR",
  UnknownError = "UNKNOWN_ERROR",
  VerificationTokenExpiredError = "VERIFICATION_TOKEN_EXPIRED_ERROR",
  VerificationTokenInvalidError = "VERIFICATION_TOKEN_INVALID_ERROR",
}

export type ErrorResult = {
  errorCode: ErrorCode
  message: Scalars["String"]["output"]
}
// ...
export type InvalidCredentialsError = ErrorResult & {
  __typename?: "InvalidCredentialsError"
  authenticationError: Scalars["String"]["output"]
  errorCode: ErrorCode
  message: Scalars["String"]["output"]
}
// ...
export type NativeAuthStrategyError = ErrorResult & {
  __typename?: "NativeAuthStrategyError"
  errorCode: ErrorCode
  message: Scalars["String"]["output"]
}
// ...
export type NotVerifiedError = ErrorResult & {
  __typename?: "NotVerifiedError"
  errorCode: ErrorCode
  message: Scalars["String"]["output"]
}
// ...
export type LoginMutationVariables = Exact<{
  username: Scalars["String"]["input"]
  password: Scalars["String"]["input"]
}>

export type LoginMutation = {
  __typename?: "Mutation"
  login:
    | { __typename?: "CurrentUser"; id: string; identifier: string }
    | {
        __typename?: "InvalidCredentialsError"
        errorCode: ErrorCode
        message: string
        authenticationError: string
      }
    | {
        __typename?: "NativeAuthStrategyError"
        errorCode: ErrorCode
        message: string
      }
    | { __typename?: "NotVerifiedError"; errorCode: ErrorCode; message: string }
}
// ...
export const LoginDocument = {
  kind: "Document",
  definitions: [
    {
      kind: "OperationDefinition",
      operation: "mutation",
      name: { kind: "Name", value: "Login" },
      variableDefinitions: [
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "username" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
        {
          kind: "VariableDefinition",
          variable: {
            kind: "Variable",
            name: { kind: "Name", value: "password" },
          },
          type: {
            kind: "NonNullType",
            type: {
              kind: "NamedType",
              name: { kind: "Name", value: "String" },
            },
          },
        },
      ],
      selectionSet: {
        kind: "SelectionSet",
        selections: [
          {
            kind: "Field",
            name: { kind: "Name", value: "login" },
            arguments: [
              {
                kind: "Argument",
                name: { kind: "Name", value: "username" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "username" },
                },
              },
              {
                kind: "Argument",
                name: { kind: "Name", value: "password" },
                value: {
                  kind: "Variable",
                  name: { kind: "Name", value: "password" },
                },
              },
            ],
            selectionSet: {
              kind: "SelectionSet",
              selections: [
                {
                  kind: "InlineFragment",
                  typeCondition: {
                    kind: "NamedType",
                    name: { kind: "Name", value: "CurrentUser" },
                  },
                  selectionSet: {
                    kind: "SelectionSet",
                    selections: [
                      { kind: "Field", name: { kind: "Name", value: "id" } },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "identifier" },
                      },
                    ],
                  },
                },
                {
                  kind: "InlineFragment",
                  typeCondition: {
                    kind: "NamedType",
                    name: { kind: "Name", value: "InvalidCredentialsError" },
                  },
                  selectionSet: {
                    kind: "SelectionSet",
                    selections: [
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "errorCode" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "message" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "authenticationError" },
                      },
                    ],
                  },
                },
                {
                  kind: "InlineFragment",
                  typeCondition: {
                    kind: "NamedType",
                    name: { kind: "Name", value: "NotVerifiedError" },
                  },
                  selectionSet: {
                    kind: "SelectionSet",
                    selections: [
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "errorCode" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "message" },
                      },
                    ],
                  },
                },
                {
                  kind: "InlineFragment",
                  typeCondition: {
                    kind: "NamedType",
                    name: { kind: "Name", value: "NativeAuthStrategyError" },
                  },
                  selectionSet: {
                    kind: "SelectionSet",
                    selections: [
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "errorCode" },
                      },
                      {
                        kind: "Field",
                        name: { kind: "Name", value: "message" },
                      },
                    ],
                  },
                },
              ],
            },
          },
        ],
      },
    },
  ],
} as unknown as DocumentNode<LoginMutation, LoginMutationVariables>